Fix compiler warnings: unused parameter and sign comparison
- Added (void)p cast to suppress unused parameter warning in ScanGapsThread
- Cast type to uint32_t in Backup.cpp to fix sign comparison warning with HEADW

+++ Backup.cpp
@@ -417,7 +417,7 @@ bool Kangaroo::SaveHeader(string fileName,FILE* f,int type,uint64_t totalCount,d
   }
   ::fwrite(&version,sizeof(uint32_t),1,f);
 
-  if(type==HEADW) {
+  if((uint32_t)type==HEADW) {
 
     // Save global param
     ::fwrite(&dpSize,sizeof(uint32_t),1,f);
